refactor(frontend): Update SnippetCreationForm and SnippetRetrievalForm to use BaseForm
Refactor both forms to extend BaseForm component
Create individual CSS modules for each form
Move specific styles and functionality to respective components

This change improves code reusability and maintainability.
