+++ frontend/src/app/components/BaseForm.module.css
@@ -27,7 +27,7 @@
 
 .button {
   padding: 0.5rem 1rem;
-  background-color: #0070f3;
+  background-color: #363a3f;
   color: white;
   border: none;
   border-radius: 4px;
@@ -37,7 +37,7 @@
 }
 
 .button:hover {
-  background-color: #0051bb;
+  background-color: #5f6a77;
 }
 
 .errorContainer {
