+++ frontend/src/app/components/SnippetCreationForm.module.css
@@ -1,0 +1,11 @@
+.container {
+  width: 100%;
+  max-width: 500px;
+  margin: 0 auto;
+}
+
+.success {
+  color: #4caf50;
+  font-size: 0.875rem;
+  margin-top: 0.5rem;
+}
